Refactor ModuleRecoveryAlert component, spliting out Actions and Description.
